Split talent.js into smaller files, remove power orders from spell upcasting (and vice-versa)
